+++ packages/modules/atlas-core/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   We added Language Selector widget to Atlas_Default and Atlas_TopBar responsive layouts.
 -   Minimal required version of Studio Pro is v9.12.0 and above.
+-   We added error style to datepicker, dropdown and referenceSelector widgets.
 
 ## [3.5.0] Atlas Core - 2022-7-15